Troubleshot :imagesdir:
Tried including :imagesdir: attribute in the partial rather than the full build .adoc.

Again, Asciidoctor and Asciidoctor-pdf handle the relative path differently (relative to the build .adoc vs relative to the included .adoc). As with my previous troubleshooting exercise in wap-doc-test-build, this requires conditionals to manage (based on backends for outputs). In the scenario for this commit, it's get tricky to manage :imagedir: if publishing the partial .adoc alone.